+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 *   New Features
     *   Show episode summaries where possible
         ([#5276](https://github.com/Automattic/pocket-casts-android/pull/5276))
+    *   Show episode summary on fullscreen player
+        ([#5278](https://github.com/Automattic/pocket-casts-android/pull/5278))
 
 8.12
 -----